requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
CacheControl==0.12.6
cachetools==4.2.2
cachy==0.3.0
Cerberus==1.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cleo==0.7.6
click==8.0.1
clikit==0.4.3
cryptography==3.4.7
cs50==6.0.4
cycler==0.10.0
debugpy==1.2.1
dephell==0.8.3
dephell-archive==0.1.7
dephell-argparse==0.1.3
dephell-changelogs==0.0.1
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-setuptools==0.2.4
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.18
dephell-versioning==0.1.2
docopt==0.6.2
docutils==0.17.1
Flask==2.0.1
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
glfw==2.1.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.0
h5py==2.10.0
html5lib==1.1
idna==2.10
importlib-metadata==3.10.1
itsdangerous==2.0.1
jedi==0.15.2
jeepney==0.6.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==20.0.1
kiwisolver==1.3.1
lockfile==0.12.2
m2r==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
nltk==3.6.2
numpy==1.20.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
Pillow==8.2.0
pipreqs-amasad==0.4.10
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.0.5
protobuf==3.17.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyflakes==2.1.1
pygame==2.0.1
pylev==1.3.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyquil==1.9.0
pyrsistent==0.14.11
python-dateutil==2.8.1
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
pyxel==1.0.1
referenceqvm==0.3
regex==2021.4.4
replit==3.1.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rope==0.18.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scipy==1.4.1
SecretStorage==3.3.1
selenium==3.141.0
shellingham==1.4.0
six==1.16.0
sounddevice==0.4.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
tomlkit==0.5.11
tqdm==4.61.0
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
yarg==0.1.9
yarl==1.6.3
yaspin==2.0.0
zipp==3.4.1
absl-py==0.12.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
APScheduler==3.7.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
CacheControl==0.12.6
cachetools==4.2.2
cachy==0.3.0
Cerberus==1.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cleo==0.7.6
click==8.0.1
clikit==0.4.3
cryptography==3.4.7
cs50==6.0.4
cycler==0.10.0
debugpy==1.2.1
dephell==0.8.3
dephell-archive==0.1.7
dephell-argparse==0.1.3
dephell-changelogs==0.0.1
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-setuptools==0.2.4
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.18
dephell-versioning==0.1.2
docopt==0.6.2
docutils==0.17.1
Flask==2.0.1
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
glfw==2.1.0
google-auth==1.30.2
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.0
h5py==2.10.0
html5lib==1.1
idna==2.10
importlib-metadata==3.10.1
itsdangerous==2.0.1
jedi==0.15.2
jeepney==0.6.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==20.0.1
kiwisolver==1.3.1
lockfile==0.12.2
m2r==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
nltk==3.6.2
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
Pillow==8.2.0
pipreqs-amasad==0.4.10
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.0.5
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyflakes==2.1.1
pygame==2.0.1
pylev==1.4.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyquil==1.9.0
pyrsistent==0.14.11
python-dateutil==2.8.1
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
pytz==2021.1
pyxel==1.0.1
referenceqvm==0.3
regex==2021.4.4
replit==3.1.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rope==0.18.0
rsa==4.7.2
ruamel.yaml==0.17.8
ruamel.yaml.clib==0.2.2
scipy==1.4.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sounddevice==0.4.1
SQLAlchemy==1.4.17
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
tomlkit==0.5.11
tqdm==4.61.0
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==1.35
urllib3==1.26.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
yarg==0.1.9
yarl==1.6.3
yaspin==2.0.0
zipp==3.4.1
